backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-multipart
python-dotenv
pydantic
requests
pypdf
